+++ src/Rib/Shake.hs
@@ -16,6 +16,7 @@ module Rib.Shake
     writeFileCached,
 
     -- * Misc
+    getCliConfig,
     ribInputDir,
     ribOutputDir,
     getDirectoryFiles',
@@ -26,26 +27,27 @@ import Development.Shake
 import Path
 import Path.IO
 import Relude
-import Rib.Settings
+import Rib.Cli (CliConfig)
+import qualified Rib.Cli as Cli
 
 -- | Get rib settings from a shake Action monad.
-ribSettings :: Action RibSettings
-ribSettings = getShakeExtra >>= \case
+getCliConfig :: Action CliConfig
+getCliConfig = getShakeExtra >>= \case
   Just v -> pure v
-  Nothing -> fail "RibSettings not initialized"
+  Nothing -> fail "CliConfig not initialized"
 
 -- | Input directory containing source files
 --
 -- This is same as the first argument to `Rib.App.run`
 ribInputDir :: Action (Path Rel Dir)
-ribInputDir = _ribSettings_inputDir <$> ribSettings
+ribInputDir = Cli.inputDir <$> getCliConfig
 
 -- | Output directory where files are generated
 --
 -- This is same as the second argument to `Rib.App.run`
 ribOutputDir :: Action (Path Rel Dir)
 ribOutputDir = do
-  output <- _ribSettings_outputDir <$> ribSettings
+  output <- Cli.outputDir <$> getCliConfig
   liftIO $ createDirIfMissing True output
   return output